Add typescript expo plugin
