+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig",
-  "exclude": ["example", "ios", "android", "e2e", "plugins"]
+  "exclude": ["example", "ios", "android", "e2e"]
 }
